Adding mixin to batch update tasks
